ajout de cast

+++ src/SFML_TextButton.cpp
@@ -19,8 +19,8 @@ SFML::TextButton::TextButton(std::string const & text,
     float const width = textWidth();
     float const height = textHeight();
     m_textButton.setFillColor(textColor);
-    assert(width <= size.x && "The width of the button must be greater than or equal to the width taken by the text");
-    assert(height <= size.y && "The height of the button must be greater than or equal to the height taken by the text");
+    assert(static_cast<int>(width) <= size.x && "The width of the button must be greater than or equal to the width taken by the text");
+    assert(static_cast<int>(height) <= size.y && "The height of the button must be greater than or equal to the height taken by the text");
 
     m_textButton.setPosition(sf::Vector2f(static_cast<float>(m_rectText.left) / 2.f - width/2.f , static_cast<float>(m_rectText.top) / 2.f - height/2.f));
     allignText();
@@ -56,7 +56,7 @@ void SFML::TextButton::setPosition(sf::Vector2i const & position) noexcept
 
 float SFML::TextButton::textWidth() const noexcept
 {
-    return m_textButton.getGlobalBounds().width/*  * m_textButton.getString().getSize() */;
+    return m_textButton.getGlobalBounds().width;
 }
 
 float SFML::TextButton::textHeight() const noexcept
